refactor: actions separated into their own file
